+++ tests/widgets/dlgbpmdetecttest.cpp
@@ -15,6 +15,8 @@ public:
     DummyBpmDetector(QObject *parent = nullptr) : AbstractBpmDetector(parent) {
     }
     void inputSamples(const soundtouch::SAMPLETYPE *samples, int numSamples) override {
+        Q_UNUSED(samples)
+        Q_UNUSED(numSamples)
     }
     bpmtype getBpm() const override {
         return 120.0;